webserver/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
pyyaml>=6.0
django-allauth[socialaccount]>=65.0
django-q2>=1.6
